Huge refactor, changed from js to jsx, zustand to recoiljs, auth modal half complete, only auth left
