+++ package.json
@@ -32,13 +32,14 @@
     "react-dom": "18.2.0",
     "react-firebase-hooks": "^5.1.1",
     "react-icons": "^4.10.1",
-    "reactfire": "^4.2.3",
+    "recoil": "^0.7.7",
     "remark": "^14.0.3",
     "remark-html": "^15.0.2",
     "tls": "^0.0.1",
     "zustand": "^4.3.8"
   },
   "devDependencies": {
+    "@types/react": "^18.2.14",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "postcss-import": "^15.1.0",
