--- pages/posts/posts.js
@@ -1,7 +1,0 @@
-function PostViewPage() {
-  return (
-    <div className="text-4xl text-inherit">Posts made by you and others...</div>
-  );
-}
-
-export default PostViewPage;
