+++ src/atoms/AuthModalAtom.jsx
@@ -1,0 +1,9 @@
+import { atom } from "recoil";
+
+export const authModalState = atom({
+  key: "authModalState",
+  default: {
+    open: false,
+    view: "signin",
+  },
+});
