+++ src/pages/api/stores.js
@@ -1,11 +1,6 @@
 import React from "react";
 import { create } from "zustand";
 
-export const useToggleStore = create((set) => ({
-  isModalOpen: 0,
-  toggleModalOpen: () => set((state) => ({ isModalOpen: !state.isModalOpen })),
-}));
-
 export const useThemeStore = create((set) => ({
   theme: "light",
   setTheme: () => set((newTheme) => ({ theme: newTheme })),
@@ -19,6 +14,7 @@ export const useCredentialsStore = create((set) => ({
   userInfo: null,
   loading: true,
   userUid: "",
+  error: "",
   setUsername: (newValue) => set((state) => ({ username: newValue })),
   setEmail: (newValue) => set((state) => ({ email: newValue })),
   setPassword: (newValue) => set((state) => ({ password: newValue })),
@@ -26,4 +22,5 @@ export const useCredentialsStore = create((set) => ({
   setUserInfo: (newValue) => set((state) => ({ userInfo: newValue })),
   setLoading: (newValue) => set((state) => ({ loading: newValue })),
   setUserUid: (newValue) => set((state) => ({ userUid: newValue })),
+  setError: (newValue) => set((state) => ({ error: newValue })),
 }));
