--- styles/globals.css
@@ -1,19 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-html,
-body {
-  @apply leading-[1.6] text-lg m-0 p-0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-}
-* {
-  @apply box-border;
-}
-a {
-  @apply text-[#0070f3] no-underline hover:underline;
-}
-img {
-  @apply max-w-full block;
-}
