--- styles/utils.module.css
@@ -1,44 +1,0 @@
-.heading2Xl {
-  @apply text-[2.5rem] leading-[1.2] font-extrabold tracking-[-0.05rem] mx-0 my-4;
-}
-.headingXl {
-  @apply text-[2rem] leading-[1.3] font-extrabold tracking-[-0.05rem] mx-0 my-4;
-}
-.headingLg {
-  @apply text-2xl leading-[1.4] mx-0 my-4;
-}
-.headingMd {
-  @apply text-[1.2rem] leading-normal;
-}
-.borderCircle {
-  @apply rounded-full;
-}
-.colorInherit {
-  @apply text-inherit;
-}
-.padding1px {
-  @apply pt-px;
-}
-.list {
-  @apply m-0 p-0;
-  list-style: none;
-}
-.listItem {
-  @apply mt-0 mb-5 mx-0;
-}
-.lightText {
-  @apply text-[#666];
-}
-.faGoogleColored {
-  @apply bg-clip-text text-transparent;
-  background: conic-gradient(
-      from -45deg,
-      #ea4335 110deg,
-      #4285f4 90deg 180deg,
-      #34a853 180deg 270deg,
-      #fbbc05 270deg
-    )
-    73% 55%/150% 150% no-repeat;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
